% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_tbl.R
\name{cat_tbl}
\alias{cat_tbl}
\title{Summarize a categorical variable}
\usage{
cat_tbl(data, var, na.rm = FALSE, only = NULL, ignore = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{var}{A character string of the name of a variable in `data` containing categorical
data.}

\item{na.rm}{A logical value indicating whether missing values should be removed
before calculations. Default is `FALSE`.}

\item{only}{A character string, or vector of character strings, of the types of
summary data to return. Default is `NULL`, which returns both counts and
percentages. To return only counts or percentages, use `count` or
`percent`, respectively.}

\item{ignore}{An optional vector that contains values or labels you wish to exclude
from the data and any following calculations. Default is `NULL`}
}
\value{
A tibble displaying the relative frequency counts and/or percentages of `var`.
}
\description{
`cat_group_tbl()` presents frequency counts and percentages (count, percent)
for nominal or categorical variables. Missing data can be excluded from the calculations.
}
\examples{
cat_tbl(data = nlsy, var = "gender")

cat_tbl(data = nlsy, var = "race", only = "count")

cat_tbl(data = nlsy,
        var = "race",
        ignore = "Hispanic",
        only = "percent",
        na.rm = TRUE)

}
\author{
Ama Nyame-Mensah
}
