% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_named_obj.R
\name{check_named_obj}
\alias{check_named_obj}
\title{Check a named vector}
\usage{
check_named_obj(x, cols, default)
}
\arguments{
\item{x}{A named list or vector.}

\item{cols}{A character vector containing column names or variable stems to match.}

\item{default}{Default value to return}
}
\value{
Either the original object or some default value.
}
\description{
This experimental function assesses whether named lists and vectors have
invalid values (like `NULL` or `NA`), invalid names (such as missing or empty names),
confirms that the count of valid names matches the count of provided values, and verifies
that the valid names obtained from the named object align with the supplied names.
}
