% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_group_tbl.R
\name{mean_group_tbl}
\alias{mean_group_tbl}
\title{Summarize continuous variables by group}
\usage{
mean_group_tbl(
  data,
  var_stem,
  escape_stem = FALSE,
  ignore_stem_case = FALSE,
  group,
  group_type = "variable",
  group_name = NULL,
  escape_group = FALSE,
  ignore_group_case = FALSE,
  remove_group_non_alnum = TRUE,
  na_removal = "listwise",
  only = NULL,
  var_labels = NULL,
  ignore = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{var_stem}{A character string of a variable stem or the full name of a variable in `data`.}

\item{escape_stem}{A logical value indicating whether to escape `var_stem`. Default is
`FALSE`.}

\item{ignore_stem_case}{A logical value indicating whether the search for columns matching
the supplied `var_stem` is case-insensitive. Default is `FALSE`.}

\item{group}{A character string of a variable in `data` or a pattern to use to search for variables
in `data`.}

\item{group_type}{A character string that defines the type of grouping variable. Should be one of
`pattern` or `variable`. Default is `variable`, in which case the variable matching the `group`
string will be searched for within `data`.}

\item{group_name}{A character string piped to the final table to replace the name of `group`.}

\item{escape_group}{A logical value indicating whether to escape string supplied to `group`.}

\item{ignore_group_case}{A logical value indicating whether `group` is case-insensitive.
Default is `FALSE`.}

\item{remove_group_non_alnum}{A logical value indicating whether to remove all non-alphanumeric
characters (anything that is not a letter or number) from `group`. Default is `TRUE`.}

\item{na_removal}{A character string specifying how to remove missing values. Should be
one of `pairwise` or `listwise`. Default is `listwise`.}

\item{only}{A character string or vector of character strings of the types of
summary data to return. Default is `NULL`, which returns mean (mean), standard
deviation (sd), minimum value (min), maximum value (max), and non-missing responses
(nobs).}

\item{var_labels}{An optional named character vector or list where each element maps
labels to variable names. If any element is unnamed or if any labels do not match variables
in returned from `data`, all labels will be ignored and the table will be printed without them.}

\item{ignore}{An optional named character vector or list specifying values to exclude from
the dataset and analysis. By default, `NULL` includes all available values. To omit values
from variables returned by `var_stem`, use the provided stem as the name. To exclude values
from both `var_stem` variables and a grouping variable in `data`, supply a list.}
}
\value{
A table presenting summary statistics for series of continuous variables with
the same variable stem, categorized by a grouping variable.
}
\description{
`mean_group_tbl()` presents descriptive statistics (mean, sd, minimum, maximum, number
of non-missing observations) for interval (e.g., Test scores) and ratio level (e.g., Age)
variables with the same variable stem by some grouping variable. A variable stem is a common prefix
found in related variable names, often corresponding to similar survey items, that represents a
shared concept before unique identifiers (like timepoints) are added. For example, in the
`stem_social_psych` dataset, the two variables 'belong_belongStem_w1' and 'belong_belongStem_w2'
share the variable stem 'belong_belongStem' (e.g., "I feel like an outsider in STEM"), with suffixes
(_w1, _w2) indicating different measurement waves. By default, missing data are excluded from the
calculations in a listwise fashion.
}
\examples{
mean_group_tbl(data = stem_social_psych,
               var_stem = "belong_welcomedStem",
               group = "_w\\\\d",
               group_type = "pattern",
               na_removal = "pairwise",
               var_labels = c("belong_welcomedStem_w1" = "I feel welcomed in STEM workplaces",
                              "belong_welcomedStem_w2" = "I feel welcomed in STEM workplaces"),
               group_name = "wave")

mean_group_tbl(data = stem_social_psych,
               var_stem = "belong",
               group = "_w\\\\d",
               group_type = "pattern",
               na_removal = "listwise",
               var_labels = c(belong_belongStem_w1 = "I feel like I belong in STEM",
                              belong_belongStem_w2 = "I feel like I belong in STEM",
                              belong_outsiderStem_w1 = "I feel like an outsider in STEM",
                              belong_outsiderStem_w2 = "I feel like an outsider in STEM",
                              belong_welcomedStem_w1 = "I feel welcomed in STEM workplaces",
                              belong_welcomedStem_w2 = "I feel welcomed in STEM workplaces"),
               group_name = "wave")

grouped_data <-
  data.frame(
    symptoms.t1 = sample(c(1:5, -999), replace = TRUE, size = 50),
    symptoms.t2 = sample(c(NA, 1:5, -999), replace = TRUE, size = 50)
  )

mean_group_tbl(data = grouped_data,
               var_stem = "symptoms",
               group = ".t\\\\d",
               group_type = "pattern",
               escape_group = TRUE,
               na_removal = "listwise",
               ignore = c(symptoms = -999))

}
\author{
Ama Nyame-Mensah
}
