% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_attach_data.R
\docType{data}
\name{peer_attach}
\alias{peer_attach}
\title{Social Psychological (Generated) Data}
\format{
A data.frame with 267 rows and 79 columns:
}
\usage{
peer_attach
}
\description{
These data analyze parents and peer attachment (congruence ) on
adolescents' self-esteem. The data were sourced from https://www.psycharchives.\\
org/en/item/9540f0f4-6d37-4b87-b6e5-29ec7d06279d
}
\keyword{datasets}
