% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_psy_data.R
\docType{data}
\name{social_psy_data}
\alias{social_psy_data}
\title{Social Psychological (Generated) Data}
\format{
A data.frame with 10,200 rows and 17 columns:
\describe{
\item{id}{participant id number)}
\item{belong_1}{I feel like I belong at this institution (1=Strongly Disagree,
2=Disagree,3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{belong_2}{I feel like part of the community (1=Strongly Disagree,
2=Disagree,3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{belong_3}{I feel valued by this institution (1=Strongly Disagree,
2=Disagree,3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{identity_1}{This institution is a big part of who I am (1=Strongly
Disagree,2=Disagree,3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{identity_2}{I feel comfortable being myself in this setting
(1=Strongly Disagree,2=Disagree,3=Neither agree nor disagree,4=Agree,
5=Strongly Agree)}
\item{identity_3}{This institution is a big part of who I am (1=Strongly Disagree,
2=Disagree,3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{identity_4}{I care about doing well at this institution (1=Strongly Disagree,
2=Disagree,3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{selfEfficacy_1}{I am confident about A (1=Strongly Disagree,2=Disagree,
3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{selfEfficacy_2}{I am confident about B (1=Strongly Disagree,2=Disagree,
3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{selfEfficacy_3}{I am confident about C (1=Strongly Disagree,2=Disagree,
3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{selfEfficacy_4}{I am confident about D (1=Strongly Disagree,2=Disagree,
3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{selfEfficacy_5}{I am confident about E (1=Strongly Disagree,2=Disagree,
3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{selfEfficacy_6}{I am confident about F (1=Strongly Disagree,2=Disagree,
3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{selfEfficacy_7}{I am confident about G (1=Strongly Disagree,2=Disagree,
3=Neither agree nor disagree,4=Agree,5=Strongly Agree)}
\item{gender}{Participant's gender identity (1=Woman,2=Man,3=Non-binary,
4=Self-identify,5=Transgender,6=Gender-queer/non-conforming)}
\item{citizen}{Participant's citizenship status (1=U.S. citizen,2=Non-U.S.
citizen with permanent residency,3=Non-U.S. citizen with temporary visa,4=Other)}
}
}
\usage{
social_psy_data
}
\description{
These data were generated to produce social psychological
data applicable to real-world contexts.
}
\keyword{datasets}
