% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytabl-package.R
\docType{package}
\name{summarytabl-package}
\alias{summarytabl}
\alias{summarytabl-package}
\title{summarytabl: Generate Summary Tables for Continuous, Ordinal, and Categorical Data}
\description{
This package creates simple summary tables for continuous, ordinal, and categorical data.
}
\author{
\strong{Maintainer}: Ama Nyame-Mensah \email{ama@anyamemensah.com}

}
\keyword{internal}
