% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_tbl.R
\name{mean_tbl}
\alias{mean_tbl}
\title{Summarize continuous variables}
\usage{
mean_tbl(
  data,
  var_stem,
  escape_stem = FALSE,
  ignore_stem_case = FALSE,
  na_removal = "listwise",
  only = NULL,
  var_labels = NULL,
  ignore = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{var_stem}{A character string of a variable stem or the full name of a variable
in \code{data}.}

\item{escape_stem}{A logical value indicating whether to escape \code{var_stem}. Default
is \code{FALSE}.}

\item{ignore_stem_case}{A logical value indicating whether the search for columns
matching the supplied \code{var_stem} is case-insensitive. Default is \code{FALSE}.}

\item{na_removal}{A character string that specifies the method for handling missing
values: \code{pairwise} or \code{listwise}. Defaults to \code{listwise}.}

\item{only}{A character string or vector of character strings specifying which summary
statistics to return. Defaults to NULL, which includes mean (mean), standard deviation
(sd), minimum (min), maximum (max), and count of non-missing values (nobs).}

\item{var_labels}{An optional named character vector or list used to assign custom
labels to variable names. Each element should be named and correspond to a variable in
the returned table. If any element is unnamed or references a variable not returned in
the table, all labels will be ignored and the table will be printed without them.}

\item{ignore}{An optional vector of values to exclude from variables identified by
\code{var_stem}. Defaults to \code{NULL}, which retains all values.}
}
\value{
A tibble showing summary statistics for continuous variables sharing a common
variable stem.
}
\description{
\code{mean_tbl()} calculates summary statistics (i.e., mean, standard
deviation, minimum, maximum, and count of non-missing values) for interval and
ratio-level variables that share a common prefix (i.e., variable stem). A variable
'stem' is a shared naming pattern across related variables, often representing
repeated measures of the same concept or a series of items measuring a single
construct. Missing data are excluded using \code{listwise} deletion by default.
}
\examples{
sdoh_child_ages <- 
  dplyr::select(sdoh, c(ACS_PCT_AGE_0_4, ACS_PCT_AGE_5_9,
                        ACS_PCT_AGE_10_14, ACS_PCT_AGE_15_17))

mean_tbl(data = sdoh_child_ages, var_stem = "ACS_PCT_AGE")

mean_tbl(data = sdoh_child_ages,
         var_stem = "ACS_PCT_AGE",
         na_removal = "pairwise",
         var_labels = c(
           ACS_PCT_AGE_0_4 = "\% of population between ages 0-4",
           ACS_PCT_AGE_5_9 = "\% of population between ages 5-9",
           ACS_PCT_AGE_10_14 = "\% of population between ages 10-14",
           ACS_PCT_AGE_15_17 = "\% of population between ages 15-17"))
                        
}
\author{
Ama Nyame-Mensah
}
