% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tbl.R
\name{select_tbl}
\alias{select_tbl}
\title{Summarize multiple response variables}
\usage{
select_tbl(
  data,
  var_stem,
  escape_stem = FALSE,
  ignore_stem_case = FALSE,
  na_removal = "listwise",
  pivot = "longer",
  only = NULL,
  var_labels = NULL,
  ignore = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{var_stem}{A character string of a variable stem or the full name of a variable
in \code{data}.}

\item{escape_stem}{A logical value indicating whether to escape \code{var_stem}. Default
is \code{FALSE}.}

\item{ignore_stem_case}{A logical value indicating whether the search for columns
matching the supplied \code{var_stem} is case-insensitive. Default is \code{FALSE}.}

\item{na_removal}{A character string that specifies the method for handling missing
values: \code{pairwise} or \code{listwise}. Defaults to \code{listwise}.}

\item{pivot}{A character string that determines the format of the table. By default,
\code{longer} returns the data in the long format. To receive the data in the \code{wide}
format, specify \code{wider}.}

\item{only}{A character string or vector of character strings of the types of
summary data to return. Default is \code{NULL}, which returns both counts and percentages.
To return only counts or percentages, use \code{count} or \code{percent}, respectively.}

\item{var_labels}{An optional named character vector or list used to assign custom
labels to variable names. Each element should be named and correspond to a variable in
the returned table. If any element is unnamed or references a variable not returned in
the table, all labels will be ignored and the table will be printed without them.}

\item{ignore}{An optional vector of values to exclude from variables identified by
\code{var_stem}. Defaults to \code{NULL}, which retains all values.}
}
\value{
A tibble showing the relative frequencies and/or percentages of multiple
response variables sharing a common variable stem.
}
\description{
\code{select_tbl()} displays frequency counts and percentages (i.e.,
count and percent) for multiple response variables, including binary variables
(such as Unselected/Selected) and ordinal variables (such as responses ranging
from strongly disagree to strongly agree), that share a common variable stem.
A variable 'stem' is a shared naming pattern across related variables, often
representing repeated measures of the same concept or a series of items measuring
a single construct. Missing data are excluded using \code{listwise} deletion by default.
}
\examples{
select_tbl(data = tas,
           var_stem = "involved_",
           na_removal = "pairwise")

select_tbl(data = depressive,
           var_stem = "dep",
           na_removal = "listwise",
           pivot = "wider",
           only = "percent")

var_label_example <-
  c("dep_1" = "how often child feels sad and blue",
    "dep_2" = "how often child feels nervous, tense, or on edge",
    "dep_3" = "how often child feels happy",
    "dep_4" = "how often child feels bored",
    "dep_5" = "how often child feels lonely",
    "dep_6" = "how often child feels tired or worn out",
    "dep_7" = "how often child feels excited about something",
    "dep_8" = "how often child feels too busy to get everything")

select_tbl(data = depressive,
           var_stem = "dep",
           na_removal = "pairwise",
           pivot = "longer",
           var_labels = var_label_example)

select_tbl(data = depressive,
           var_stem = "dep",
           na_removal = "pairwise",
           pivot = "wider",
           only = "count",
           var_labels = var_label_example)

}
\author{
Ama Nyame-Mensah
}
