% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytabl-package.R
\docType{package}
\name{summarytabl-package}
\alias{summarytabl}
\alias{summarytabl-package}
\title{summarytabl: Generate Summary Tables for Categorical, Ordinal, and Continuous Data}
\description{
Provides functions for tabulating and summarizing categorical, multiple response, ordinal, and continuous variables in R data frames. Makes it easy to create clear, structured summary tables, so you spend less time wrangling data and more time interpreting it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anyamemensah.github.io/summarytabl/}
  \item \url{https://github.com/anyamemensah/summarytabl}
  \item Report bugs at \url{https://github.com/anyamemensah/summarytabl/issues}
}

}
\author{
\strong{Maintainer}: Ama Nyame-Mensah \email{ama@anyamemensah.com}

}
\keyword{internal}
