% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytabl-package.R
\docType{package}
\name{summarytabl-package}
\alias{summarytabl}
\alias{summarytabl-package}
\title{summarytabl: Generate Summary Tables for Continuous, Ordinal, and Categorical Data}
\description{
Provides functions for tabulating and summarizing continuous, ordinal, and categorical variables in data frames. The package was designed to streamline exploratory data analysis and simplify the creation of summary tables for reports and other purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anyamemensah.github.io/summarytabl/}
}

}
\author{
\strong{Maintainer}: Ama Nyame-Mensah \email{ama@anyamemensah.com}

}
\keyword{internal}
